# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VenueFlow API - Hospitality venue management SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /time-clock/in:
    post:
      operationId: clockIn
      tags: [timeClock]
      summary: Clock in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInBody"
      responses:
        "201":
          description: Clocked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeClockEntry"
  /time-clock/out:
    post:
      operationId: clockOut
      tags: [timeClock]
      summary: Clock out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutBody"
      responses:
        "200":
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeClockEntry"
  /time-clock/active:
    get:
      operationId: listActiveClockIns
      tags: [timeClock]
      summary: List currently clocked in users
      parameters:
        - name: venueId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active clock-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeClockEntry"
  /time-clock/entries:
    get:
      operationId: listTimeClockEntries
      tags: [timeClock]
      summary: List time clock entries
      parameters:
        - name: venueId
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeClockEntry"
  /time-clock/entries/{id}:
    put:
      operationId: updateTimeClockEntry
      tags: [timeClock]
      summary: Manual edit time clock entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeClockEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeClockEntry"

  # ── Time Off ──────────────────────────────────────────────────────────────────
  /time-off:
    get:
      operationId: listTimeOffRequests
      tags: [timeOff]
      summary: List time-off requests
      parameters:
        - name: venueId
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Time-off requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeOffRequest"
    post:
      operationId: createTimeOffRequest
      tags: [timeOff]
      summary: Submit time-off request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeOffRequestBody"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeOffRequest"
  /time-off/{id}/approve:
    put:
      operationId: approveTimeOff
      tags: [timeOff]
      summary: Approve time-off request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeOffRequest"
  /time-off/{id}/deny:
    put:
      operationId: denyTimeOff
      tags: [timeOff]
      summary: Deny time-off request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeOffRequest"

  # ── Floor Plan ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TimeClockEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
          nullable: true
        venueId:
          type: string
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
          nullable: true
        totalHours:
          type: number
          nullable: true
        breakMinutes:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, completed]
        notes:
          type: string
          nullable: true
      required: [id, userId, venueId, clockIn, status]
    ClockInBody:
      type: object
      properties:
        userId:
          type: string
        venueId:
          type: string
        notes:
          type: string
      required: [userId, venueId]
    ClockOutBody:
      type: object
      properties:
        userId:
          type: string
        venueId:
          type: string
        breakMinutes:
          type: integer
        notes:
          type: string
      required: [userId, venueId]
    UpdateTimeClockEntryBody:
      type: object
      properties:
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
        breakMinutes:
          type: integer
        notes:
          type: string
    TimeOffRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        type:
          type: string
          enum: [vacation, sick, personal]
        status:
          type: string
          enum: [pending, approved, denied]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, startDate, endDate, type, status, createdAt]
    CreateTimeOffRequestBody:
      type: object
      properties:
        userId:
          type: string
        venueId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        type:
          type: string
          enum: [vacation, sick, personal]
        notes:
          type: string
      required: [userId, venueId, startDate, endDate, type]
...